+++ test/test_policies.cpp
@@ -1,4 +1,4 @@
-#define BOOST_TEST_MODULE "policy JSON parser tests"
+#define BOOST_TEST_MODULE "policy"
 
 #include "entrypoint/policy/parser.h"
 #include <boost/test/unit_test.hpp>
@@ -133,83 +133,4 @@ BOOST_AUTO_TEST_CASE(check_allow_all_policy) {
         BOOST_CHECK(result.has_value());
         BOOST_CHECK(*result == effect::allow);
     }
-}
-
-using json = nlohmann::json;
-static std::string create_test_deny_policy_with_condition(std::string_view sv) {
-    auto policy_template = json::parse(R"json({
-            "Sid":  "TestCondition",
-            "Version": "2012-10-17",
-            "Statement": {
-                "Effect": "Deny",
-                "Action": "*",
-                "Principal": "*",
-                "Resource": "*",
-                "Condition": {}
-            }
-        })json");
-    policy_template["Statement"]["Condition"] = json::parse(sv);
-
-    auto str = policy_template.dump(4);
-    return str;
-}
-
-BOOST_AUTO_TEST_CASE(check_condition__handles_lessthan_and_greaterthan) {
-    auto policy = parser::parse(create_test_deny_policy_with_condition(
-        R"json({
-            "DateGreaterThan" : {"aws:CurrentTime" : "2020-01-01T00:00:01Z"},
-            "DateLessThan" : {"aws:CurrentTime" : "2224-01-01T00:00:01Z"}
-        })json"));
-
-    auto result = policy.front().check(
-        variables(make_request("GET /test HTTP/1.1\r\n\r\n"),
-                  mock_command("s3:ListBucket")));
-
-    BOOST_CHECK(result.has_value());
-    BOOST_CHECK(*result == effect::deny);
-}
-
-BOOST_AUTO_TEST_CASE(
-    check_condition__do_not_add_deny_policy_with_equals_condition_for_different_dates) {
-    auto policy = parser::parse(create_test_deny_policy_with_condition(
-        R"json({
-            "DateEquals" : {"aws:CurrentTime" : "2011-01-01T00:00:01Z"}
-        })json"));
-
-    auto result = policy.front().check(
-        variables(make_request("GET /test HTTP/1.1\r\n\r\n"),
-                  mock_command("s3:ListBucket")));
-
-    BOOST_CHECK(!result.has_value());
-}
-
-BOOST_AUTO_TEST_CASE(
-    check_condition__adds_deny_policy_with_notequals_condition_for_different_dates) {
-    auto policy = parser::parse(create_test_deny_policy_with_condition(
-        R"json({
-            "DateNotEquals" : {"aws:CurrentTime" : "2011-01-01T00:00:01Z"}
-        })json"));
-
-    auto result = policy.front().check(
-        variables(make_request("GET /test HTTP/1.1\r\n\r\n"),
-                  mock_command("s3:ListBucket")));
-
-    BOOST_CHECK(result.has_value());
-    BOOST_CHECK(*result == effect::deny);
-}
-
-BOOST_AUTO_TEST_CASE(check_condition__does_not_handle_list_type_value) {
-    auto policy = parser::parse(create_test_deny_policy_with_condition(
-        R"json({
-            "DateGreaterThan": {
-                "aws:CurrentTime": [
-                    "2020-01-01T00:00:01Z", "2020-01-01T00:00:01Z"
-                ]
-            }
-        })json"));
-
-    BOOST_REQUIRE_THROW(policy.front().check(variables(
-                            make_request("GET /test HTTP/1.1\r\n\r\n"),
-                            mock_command("s3:ListBucket"))),
-                        std::runtime_error);
 }